requirements.txt
# LangChain core + integrations
langchain
langchain-community
langchain-core
langchain-openai

# OpenAI client (compatible with langchain)
openai==1.79.0

# FastAPI and ASGI server
fastapi
uvicorn

# Env variable loading
python-dotenv
rapidfuzz==2.9.0
